+++ changelog.md
@@ -1,4 +1,3 @@
-- Fixed Fuel Station model
-- Updated to newest forge version
+- Fixed ejecting the player out of the car
 
 __NOTE THAT YOU HAVE TO USE AT LEAST FORGE 28.0.105__